Tested CogAgent implementation
Inference code works but the problem is that CogAgent can't ingest
multiple images at once in one forward pass due to the high resolution
cross attention module. I have some ideas on how we might be able to get
past this but not entirely sure it's worth it, let's discuss.